feat: Add user to match

+++ src/main/java/com/example/pokerpage/enums/ErrorEnum.java
@@ -4,15 +4,15 @@ import lombok.Getter;
 
 @Getter
 public enum ErrorEnum {
-    GENERAL(0L, "Erro desconhecido!"),
-    NOT_FOUND(404L, "Registro não encontrado!"),
-    MANDATORY_FIELD_NOT_FOUND(1L, "Campo Obrigatório não preenchido"),
-    DATE_START_AFTER_DATE_FINAL(2L, "A data inicial é maior que a final!");
+    GENERAL(0, "Erro desconhecido!"),
+    NOT_FOUND(404, "Registro não encontrado!"),
+    MANDATORY_FIELD_NOT_FOUND(1, "Campo Obrigatório não preenchido"),
+    DATE_START_AFTER_DATE_FINAL(2, "A data inicial é maior que a final!");
 
-    private final Long id;
+    private final Integer id;
     private final String message;
 
-    ErrorEnum(Long id, String message){
+    ErrorEnum(Integer id, String message){
         this.id = id;
         this.message = message;
     }
